+++ public/js/lib/common/login.js
@@ -20,15 +20,20 @@ export function resetCheckAuth () {
 export function setLoginState (bool, id) {
   Cookies.set('loginstate', bool, {
     expires: 365,
-    sameSite: window.cookiePolicy
+    sameSite: window.cookiePolicy,
+    secure: window.location.protocol === 'https:'
   })
   if (id) {
     Cookies.set('userid', id, {
       expires: 365,
-      sameSite: window.cookiePolicy
+      sameSite: window.cookiePolicy,
+      secure: window.location.protocol === 'https:'
     })
   } else {
-    Cookies.remove('userid')
+    Cookies.remove('userid', {
+      sameSite: window.cookiePolicy,
+      secure: window.location.protocol === 'https:'
+    })
   }
   lastLoginState = bool
   lastUserId = id