+++ public/js/lib/editor/index.js
@@ -343,13 +343,15 @@ export default class Editor {
       if (this.editor.getOption('indentWithTabs')) {
         Cookies.set('indent_type', 'tab', {
           expires: 365,
-          sameSite: window.cookiePolicy
+          sameSite: window.cookiePolicy,
+          secure: window.location.protocol === 'https:'
         })
         type.text('Tab Size:')
       } else {
         Cookies.set('indent_type', 'space', {
           expires: 365,
-          sameSite: window.cookiePolicy
+          sameSite: window.cookiePolicy,
+          secure: window.location.protocol === 'https:'
         })
         type.text('Spaces:')
       }
@@ -361,12 +363,14 @@ export default class Editor {
       if (this.editor.getOption('indentWithTabs')) {
         Cookies.set('tab_size', unit, {
           expires: 365,
-          sameSite: window.cookiePolicy
+          sameSite: window.cookiePolicy,
+          secure: window.location.protocol === 'https:'
         })
       } else {
         Cookies.set('space_units', unit, {
           expires: 365,
-          sameSite: window.cookiePolicy
+          sameSite: window.cookiePolicy,
+          secure: window.location.protocol === 'https:'
         })
       }
       widthLabel.text(unit)
@@ -435,7 +439,8 @@ export default class Editor {
       const keymap = this.editor.getOption('keyMap')
       Cookies.set('keymap', keymap, {
         expires: 365,
-        sameSite: window.cookiePolicy
+        sameSite: window.cookiePolicy,
+        secure: window.location.protocol === 'https:'
       })
       label.text(keymap)
       this.restoreOverrideEditorKeymap()
@@ -484,7 +489,8 @@ export default class Editor {
       this.editor.setOption('theme', theme)
       Cookies.set('theme', theme, {
         expires: 365,
-        sameSite: window.cookiePolicy
+        sameSite: window.cookiePolicy,
+        secure: window.location.protocol === 'https:'
       })
 
       checkTheme()
@@ -530,7 +536,8 @@ export default class Editor {
       }
       Cookies.set('spellcheck', mode === 'spell-checker', {
         expires: 365,
-        sameSite: window.cookiePolicy
+        sameSite: window.cookiePolicy,
+        secure: window.location.protocol === 'https:'
       })
 
       checkSpellcheck()
@@ -577,7 +584,8 @@ export default class Editor {
     if (overrideBrowserKeymap.is(':checked')) {
       Cookies.set('preferences-override-browser-keymap', true, {
         expires: 365,
-        sameSite: window.cookiePolicy
+        sameSite: window.cookiePolicy,
+        secure: window.location.protocol === 'https:'
       })
       this.restoreOverrideEditorKeymap()
     } else {